Rename package map to pm
