+++ src/highdicom/__init__.py
@@ -1,5 +1,5 @@
 from highdicom import legacy
-from highdicom import map
+from highdicom import pm
 from highdicom import sc
 from highdicom import seg
 from highdicom import sr
@@ -29,7 +29,7 @@ __all__ = [
     'io',
     'IssuerOfIdentifier',
     'legacy',
-    'map',
+    'pm',
     'PixelMeasuresSequence',
     'PlanePositionSequence',
     'PlaneOrientationSequence',
